Repair what the merge with main got wrong
Three things, all of them the merge rather than either side.

tests/js/gif-quantize.test.js: both branches added a file of that name - mine
for the gif-maker quantizer, video-to-gif's for its own - and the merge spliced
the two into one file, mid-import, which is a syntax error and took the whole
suite down. video-to-gif's file is restored exactly as it stands on main, and
mine moves out of the way of the name it does not own: gif-maker.test.js and
gif-maker-quantize.test.js, which is the tests/README.md convention and says
which tool each file belongs to now that there are two GIF tools.

config/planned.toml: both sides deleted a line from the "GIF & animation" group
- main removed "Video to GIF" when it shipped, this branch removed "GIF maker" -
and the merge kept both. The roadmap would have listed two tools as planned
while linking to both of them as built. Both lines go.

tools/README.md is generated, and the merge left its order out of step with the
hub order in config/site.toml. Rebuilt.

927 JS tests and 295 Python tests pass, the build writes 68 pages, and the
output is still what Chromium's ImageDecoder says it is: three frames, 100/200/
300 ms, looping forever.
