+++ config/planned.toml
@@ -34,8 +34,6 @@ items = [
 id = "gif"
 name = "GIF &amp; animation"
 items = [
-  ["Video to GIF", "pick the section, the size, and the frame rate"],
-  ["GIF maker", "images into an animated GIF"],
   ["GIF to MP4 or WebM", "the same animation, usually a tenth of the size"],
   ["Split a GIF", "every frame out as its own PNG"],
   ["Resize, crop &amp; rotate GIFs", "frame by frame, with the timing kept"],
