+++ tests/js/gif-maker.test.js
